plotmyvert_backend/pyproject.toml
[tool.poetry]
name = "plotmyvert-backend"
version = "0.1.0"
description = ""
authors = ["Steven Duzevich <steven@duz.ie>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
djangorestframework = "^3.14.0"
pandas = "^2.1.3"
lxml = "^4.9.3"
numpy = "^1.26.2"
plotly = "^5.18.0"
django-cors-headers = "^4.3.1"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
